backend/requirements.txt
# Backend ETL Multi-Tenant - ConectaBoi Insight
# Dependências para o servidor Flask

# Framework web
Flask==3.0.0
Flask-CORS==4.0.0

# Autenticação e JWT
PyJWT>=2.10.1,<3.0.0

# Supabase
supabase==2.17.0

# Processamento de dados
pandas>=2.0.0
numpy>=1.24.0

# Utilitários
requests>=2.31.0
python-dotenv==1.0.0

# Para arquivos Excel
openpyxl>=3.1.0
xlrd>=2.0.0

# Para desenvolvimento
python-dateutil>=2.8.2 